+++ pentobi/AppearanceDialog.qml
@@ -11,7 +11,6 @@ import QtQuick.Controls
 PentobiDialog {
     id: root
 
-    property int currentThemeIndex
     property int currentMoveMarkingIndex
     property int currentCommentIndex
 
@@ -19,17 +18,7 @@ PentobiDialog {
     onAboutToShow: {
         switchCoordinates.checked = gameView.showCoordinates
         switchShowVariations.checked = gameModel.showVariations
-        if (themeName === "Dark")
-            currentThemeIndex = 1
-        else if (themeName === "ColorblindLight")
-            currentThemeIndex = 2
-        else if (themeName === "ColorblindDark")
-            currentThemeIndex = 3
-        else if (themeName === "System")
-            currentThemeIndex = isAndroid ? 1 : 4
-        else
-            currentThemeIndex = 0
-        comboBoxTheme.currentIndex = currentThemeIndex
+        switchColorblind.checked = theme.colorblind
         if (gameView.moveMarking === "last_dot")
             currentMoveMarkingIndex = 0
         else if (gameView.moveMarking === "last_number")
@@ -54,13 +43,7 @@ PentobiDialog {
     onAccepted: {
         gameView.showCoordinates = switchCoordinates.checked
         gameModel.showVariations = switchShowVariations.checked
-        switch (comboBoxTheme.currentIndex) {
-        case 0: themeName = "Light"; break
-        case 1: themeName = "Dark"; break
-        case 2: themeName = "ColorblindLight"; break
-        case 3: themeName = "ColorblindDark"; break
-        case 4: themeName = "System"; break
-        }
+        theme.colorblind = switchColorblind.checked
         switch (comboBoxMoveMarking.currentIndex) {
         case 0: gameView.moveMarking = "last_dot"; break
         case 1: gameView.moveMarking = "last_number"; break
@@ -110,37 +93,15 @@ PentobiDialog {
                 text: qsTr("Move number")
                 checked: gameView.showMoveNumber
             }
-            Label {
-                text: qsTr("Color theme:")
-                Layout.topMargin: 0.5 * font.pixelSize
+            Switch {
+                id: switchColorblind
 
-
-            }
-            ComboBox {
-                id: comboBoxTheme
-
-                model: isAndroid ?
-                           [
-                               qsTr("Light"),
-                               qsTr("Dark"),
-                               qsTr("Colorblind light"),
-                               qsTr("Colorblind dark")
-                           ] :
-                           [
-                               qsTr("Light"),
-                               qsTr("Dark"),
-                               qsTr("Colorblind light"),
-                               qsTr("Colorblind dark"),
-                               //: Name of theme using default system colors
-                               qsTr("System")
-                           ]
-                Layout.preferredWidth: font.pixelSize * 20
-                Layout.fillWidth: true
+                text: qsTr("Colorblind")
+                checked: rootWindow.theme.colorblind
             }
             Label {
                 text: qsTr("Move marking:")
                 Layout.topMargin: 0.5 * font.pixelSize
-
 
             }
             ComboBox {